backend-engineer/microservices/notification_service/requirements.txt
redis
pydantic
python-dotenv
